+++ test/material/binarycoefficients/test_binarycoefficients.cc
@@ -38,6 +38,7 @@
 #include <dumux/material/binarycoefficients/h2o_xylene.hh>
 #include <dumux/material/binarycoefficients/n2_o2.hh>
 
+#include <dumux/material/components/co2.hh>
 #include <test/porousmediumflow/co2/co2tables.hh>
 
 template<class Scalar, class BinaryCoefficients>
@@ -93,7 +94,7 @@ int main()
 
     success += checkBinaryCoefficients<Scalar, Air_Mesitylene>();
     success += checkBinaryCoefficients<Scalar, Air_Xylene>();
-    success += checkBinaryCoefficients<Scalar, Brine_CO2<Scalar, GeneratedCO2Tables::CO2Tables>>();
+    success += checkBinaryCoefficients<Scalar, Brine_CO2<Scalar, Components::CO2<Scalar, GeneratedCO2Tables::CO2Tables>>>();
     success += checkBinaryCoefficients<Scalar, H2O_Air>();
     success += checkBinaryCoefficients<Scalar, H2O_CH4>();
     success += checkBinaryCoefficients<Scalar, H2O_Component<Scalar, Components::Constant<0, Scalar>>>();